backend/requirements.txt
fastapi
uvicorn[standard]
starlette
pydantic
pydantic-settings
python-multipart

# Video/Audio processing
ffmpeg-python
opencv-python
librosa
soundfile
scipy
numpy
scikit-learn
soxr

# Scene detection
scenedetect[opencv]

# Transcription
openai-whisper

# HTTP
requests
